requirements.txt
openai==1.78.1
python-dotenv==1.1.0
numpy==1.26.4
httpx>=0.24.1
tqdm>=4.66.1
distro>=1.8.0
